requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
blinker==1.9.0
braceexpand==0.1.7
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
clip @ git+https://github.com/openai/CLIP.git@ded190a052fdf4585bd685cee5bc96e0310d2c93
colorama==0.4.6
datasets==4.5.0
decorator==5.2.1
dill==0.4.0
ffmpeg-python==0.2.0
filelock==3.20.3
flake8==7.3.0
Flask==3.1.2
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h5py==3.15.1
hdbscan==0.8.41
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
laion_clap==1.1.7
lazy_loader==0.4
librosa==0.11.0
librt==0.8.1
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.3
numba==0.64.0
numpy==1.26.4
open_clip_torch==3.3.0
opencv-python==4.8.1.78
packaging==26.0
pandas==3.0.0
pathspec==1.0.4
pillow==12.0.0
platformdirs==4.9.2
pluggy==1.6.0
pooch==1.9.0
progressbar==2.5
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.0
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.25
tokenizers==0.22.2
torch==2.10.0
torchlibrosa==0.1.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
typer==0.23.0
typer-slim==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.25.0
wcwidth==0.6.0
webdataset==1.0.2
Werkzeug==3.1.3
wget==3.2
xxhash==3.6.0
yarl==1.22.0
